Created Encapsulation for Reading Text Files, Creating Tree, and Applying Heuristic Function
KRUTIK - There are now two classes for reading the enemy teams text file
in our ChessAI.py file. These functions are on line 341 and 348 named
Def_Enemy and Off_Enemy. You need to be able to read from the
corresponding text file and turn that into a moveTuple object to be
passed to main. In main, the moveTuple will be used in the function
MakeMove.
KEVIN - I moved the Tree() class into chess AI.py. If you can make it to
work recursively, we will then create the tree inside the GetMove
functions in both the Off_Heuristic and Def_Heuristic classes. Then we
apply the heuristic functions and return the moveTuple.
Eric - Create the stalemate condition within the top of the main loop in
PythonChessMain.py

We should be able to piece the program together pretty easily with all
of that finished. Try to get as much done as you can before we meet
tomorrow.